#12 Changes to directory structure, allow dash pages, created template for airport analysis
