+++ pages/Airports_Analysis.py
@@ -1,0 +1,87 @@
+import polars as pl
+from dash import html, dcc, Input, Output, State
+import dash_bootstrap_components as dbc
+import plotly_express as px
+
+import dash
+
+import flask
+
+from .flightsNavbar import navbar_named
+from .AirportAnalyticsText import airport_text
+
+dash.register_page(__name__, path='/AirportAnalytics')
+
+textResults = airport_text()
+
+
+layout = html.Div([dbc.Container([
+    dbc.Row([
+        dbc.Col([
+            navbar_named('Airport Analytics')
+        ], width=12)
+    ]),
+    dbc.Row([
+
+        dbc.Col(children = textResults,
+                 id='sidebar', className = 'textSidebar',width = 12, lg = 4, md =12, xl = 4, xxl = 4, sm = 12, align='start'),
+
+        dbc.Col(children=[
+
+            dbc.Row([dbc.Col([
+
+                ## Header for the Graph Section
+                html.Hr(),
+
+                dbc.Row([
+                    dbc.Col([
+                        
+                        dbc.Card(
+                            dbc.CardBody(
+                                [
+
+                                ], className='pt-3 pb-2'
+                            ), className='bg-light text-dark border rounded-3'
+                        )
+
+                    ], width=12, md=8, id='carrier-selection-col', className='mb-2'),
+
+                    dbc.Col([
+
+                        dbc.Card(
+                            dbc.CardBody([
+
+                            ], className='pt-3 pb-2'
+                            ), className='bg-light text-dark border rounded-3'
+                        )
+
+                    ], width=12, md=4, id='visual-selection-col', className='mb-2')
+                ]),
+
+                
+
+                
+                
+
+                ## Creating Graph Component for quick analytics ##
+                dbc.Spinner([
+                    html.Div([
+                        html.H2('', id='airports-graph-header'),
+                        html.Hr(className='my-2'),
+                        html.P('', className='mb-2', id='airports-graph-desc'),
+
+                        dcc.Graph(id='airports-graph', style={'height': '50vh'})
+                    ], className='p-4 bg-light text-dark border rounded-3', id='airports_visual_div')
+                ], show_initially=True, spinner_style={'height': '3rem', 'width': '3rem'})
+
+
+                
+
+            ])], align="center")
+            
+            
+
+        ], id='airports-graph_section', width = 12, lg = 8, md = 12, xl = 8, xxl = 8, sm=12, class_name='px-2')
+
+    ])
+], fluid=True, className='mr-4 ml-4')], style={'margin': 0})
