+++ pages/Passenger_Utilization_Summary.py
@@ -65,5 +65,3 @@ summary = dag.AgGrid(
 
 
 
-
-
